code refactoring: moved api call logic to container
