+++ app/containers/App/constants.js
@@ -10,9 +10,9 @@
  */
 
 export const DEFAULT_LOCALE = 'en';
-export const LOAD_REPOS = 'boilerplate/App/LOAD_REPOS';
-export const LOAD_REPOS_SUCCESS = 'boilerplate/App/LOAD_REPOS_SUCCESS';
-export const LOAD_REPOS_ERROR = 'boilerplate/App/LOAD_REPOS_ERROR';
-export const LOAD_DATA = 'app/ApiPage/LOAD_DATA';
-export const LOAD_DATA_SUCCESS = 'app/ApiPage/LOAD_DATA_SUCCESS';
-export const LOAD_DATA_ERROR = 'app/ApiPage/LOAD_DATA_ERROR';
+// export const LOAD_REPOS = 'boilerplate/App/LOAD_REPOS';
+// export const LOAD_REPOS_SUCCESS = 'boilerplate/App/LOAD_REPOS_SUCCESS';
+// export const LOAD_REPOS_ERROR = 'boilerplate/App/LOAD_REPOS_ERROR';
+// export const LOAD_DATA = 'app/ApiPage/LOAD_DATA';
+// export const LOAD_DATA_SUCCESS = 'app/ApiPage/LOAD_DATA_SUCCESS';
+// export const LOAD_DATA_ERROR = 'app/ApiPage/LOAD_DATA_ERROR';
