+++ app/containers/App/reducer.js
@@ -12,12 +12,7 @@
 
 import { fromJS } from 'immutable';
 
-import {
-  LOAD_DATA_SUCCESS,
-  LOAD_DATA,
-  LOAD_DATA_ERROR,
 
-} from './constants';
 
 // The initial state of the App
 const initialState = fromJS({
@@ -31,25 +26,9 @@ const initialState = fromJS({
 });
 
 function appReducer(state = initialState, action) {
-  switch (action.type) {
-    case LOAD_DATA:
-      return state
-        .set('loading', true)
-        .set('error', false)
-        .setIn('data', false);
-    case LOAD_DATA_SUCCESS:
-      return state
-        // .setIn('data', action.DATA)
-        .set('data', action.DATA)
-        .set('loading', false);
-        // .set('currentUser', action.username);
-    case LOAD_DATA_ERROR:
-      return state
-        .set('error', action.error)
-        .set('loading', false);
-    default:
+  
       return state;
-  }
+  
 }
 
 export default appReducer;
