+++ app/containers/App/selectors.js
@@ -2,34 +2,8 @@
  * The global state selectors
  */
 
-import { createSelector } from 'reselect';
 
 const selectGlobal = (state) => state.get('global');
-const makeSelectCurrentUser = () => createSelector(
-  selectGlobal,
-  (globalState) => globalState.get('currentUser')
-);
-
-const makeSelectLoading = () => createSelector(
-  selectGlobal,
-  (globalState) => globalState.get('loading')
-);
-
-const makeSelectError = () => createSelector(
-  selectGlobal,
-  (globalState) => globalState.get('error')
-);
-
-const makeSelectRepos = () => createSelector(
-  selectGlobal,
-  (globalState) => globalState.getIn(['userData', 'quotes'])
-);
-
-const makeSelectData = () => createSelector(
-  selectGlobal,
-  (globalState) => globalState.getIn(['data'])
-);
-
 const makeSelectLocationState = () => {
   let prevRoutingState;
   let prevRoutingStateJS;
@@ -48,10 +22,6 @@ const makeSelectLocationState = () => {
 
 export {
   selectGlobal,
-  makeSelectCurrentUser,
-  makeSelectLoading,
-  makeSelectError,
-  makeSelectRepos,
   makeSelectLocationState,
-  makeSelectData,
+
 };
