+++ app/containers/EditQuotePage/constants.js
@@ -5,3 +5,6 @@
  */
 
 export const DEFAULT_ACTION = 'app/EditQuote/DEFAULT_ACTION';
+export const LOAD_DATA = 'app/EditQuote/LOAD_DATA';
+export const LOAD_DATA_SUCCESS = 'app/EditQuote/LOAD_DATA_SUCCESS';
+export const LOAD_DATA_ERROR = 'app/EditQuote/LOAD_DATA_ERROR';
